+++ packages/mobx-react-component/CHANGELOG.md
@@ -1,4 +1,11 @@
-# v0.50.4
+## v0.51.0
 
--   Some fixes to properly show the component name in the timings section of the performance reports in Chrome for `mobxComponent()` decorated classes
--   Added `displayName` option to the option settings of `mobxObserver`
+- Fixed SSR warnings about useLayoutEffect
+- Fixed props typings of `MobxComponent` not including children by default like class components usually do
+- Added `runBeforeMount` option to `useMobxEffects`
+- Added `useMobxStatingRendering` and `isUsingMobxStaticRendering`
+
+## v0.50.4
+
+- Some fixes to properly show the component name in the timings section of the performance reports in Chrome for `mobxComponent()` decorated classes
+- Added `displayName` option to the option settings of `mobxObserver`